+++ resources/sql/queries.sql
@@ -262,3 +262,19 @@ INSERT INTO stocks
 (login, a_id)
 VALUES
 (:login, :a_id)
+
+-- ----------------
+-- midterm section
+-- ----------------
+
+-- :name create-midterm-result! :! :n
+-- :doc create midterm grading
+INSERT INTO midterm
+(num, login, answer_id, grading)
+VALUES
+(:num, :login, :answer_id, :grading)
+
+-- :name clear-midterm! :! :n
+-- :doc clear midterm table
+DELETE from midterm
+