+++ src/clj/py99/routes/login.clj
@@ -9,7 +9,7 @@
    [struct.core :as st]
    #_[py99.db.core :as db]))
 
-(def ^:private version "0.46.0")
+(def ^:private version "0.53.2")
 
 (def ^:private l22 "https://l22.melt.kyutech.ac.jp")